update .gitignore
The return of the criba function change from int[] to arrayList to avoid using two for one for looking the size of the array and other to fill the numbers, instead one is used increasing the speed. Extract three  methods from the single function to ease code reading and undertanding.